projet fini

+++ fract_ol.h
@@ -6,7 +6,7 @@
 /*   By: yschecro <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/07 17:46:34 by yschecro          #+#    #+#             */
-/*   Updated: 2022/05/20 20:53:30 by yschecro         ###   ########.fr       */
+/*   Updated: 2022/05/24 14:46:28 by yschecro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -77,7 +77,7 @@ int			mandelbrot(int rate, complex c);
 int			julia(int rate, complex c);
 int			burning_ship(int rate, complex c);
 int			rgb(int iter,int mode, int maths);
-int			render(float len, float o_x, float o_y);
+int			render(double len, double o_x, double o_y);
 int			ft_strcheck(char *s1, char *s2);
 int			mouse_hook(int mousecode, int x, int y, void *param);
 int			close_win(int param);
